forced dtype float32 in ccdp.combine

+++ lib/class_fits_sequence.py
@@ -135,7 +135,8 @@ class FITSSequence:
             sigma_clip_func=np.ma.median,
             sigma_clip_dev_func=mad_std,
             mem_limit=600e7,
-            unit='adu'
+            unit='adu',
+            dtype='float32',
         )
 
         stack.meta['COMBINED'] = True
